Several changes and updates:  (1) Added a new keyword "required" to the tech file format "version" section.  This can be used to specify the version of magic that must be used to be compatible with the tech file. This effectively supercedes the technology version number.  (2) Changed the behavior of "make" to set the version and revision numbers on doing "make" instead of "configure".  This allows the version to update correctly after doing a "git pull" followed by "make" without doing "configure" in between.  (3) Fixed a couple of issues that were flagged as compile-time warnings.

+++ commands/CmdLQ.c
@@ -1298,7 +1298,7 @@ CmdPort(w, cmd)
     char **msg;
     int argstart;
     int i, refidx, idx, pos, type, option, argc;
-    unsigned short dirmask;
+    unsigned int dirmask;
     bool found;
     bool nonEdit = FALSE;
     Label *lab, *sl;
